[feature] 다이얼로그 파라미터 수정 및 WarmUpStopDialog 추가
