+++ presentation/src/main/java/com/dpm/sixpack/presentation/routes/session/component/dialog/WarmUpStopDialog.kt
@@ -8,26 +8,25 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.dpm.sixpack.presentation.R
-import com.dpm.sixpack.presentation.common.components.DoRunDefaultButton
+import com.dpm.sixpack.presentation.routes.session.component.panel.RecordStopButton
 
 @Composable
-internal fun WarmUpSkipDialog(
+internal fun WarmUpStopDialog(
     onCancelClick: () -> Unit,
     onStopConfirmClick: () -> Unit,
 ) {
     RunningDialog(
         onDismissRequest = {},
-        title = stringResource(R.string.dialog_warmup_skip_title),
-        contentText = stringResource(R.string.dialog_warmup_skip_content),
+        title = stringResource(R.string.dialog_warmup_stop_title),
+        contentText = stringResource(R.string.dialog_warmup_stop_content),
         onCancelClick = onCancelClick,
     ) {
-        DoRunDefaultButton(
+        RecordStopButton(
             modifier =
                 Modifier
                     .fillMaxWidth()
                     .weight(1f)
                     .height(56.dp),
-            text = stringResource(R.string.panel_skip),
             onClick = onStopConfirmClick,
         )
     }
@@ -35,8 +34,8 @@ internal fun WarmUpSkipDialog(
 
 @Preview
 @Composable
-private fun WarmUpSkipDialogPreview() {
-    WarmUpSkipDialog(
+private fun WarmUpStopDialogPreview() {
+    WarmUpStopDialog(
         onCancelClick = {},
         onStopConfirmClick = {},
     )
